feat:修复批量问题
