+++ src/main/java/cn/authing/sdk/java/dto/SendUserResetPasswordEmailLinkDto.java
@@ -1,0 +1,42 @@
+package cn.authing.sdk.java.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class SendUserResetPasswordEmailLinkDto {
+
+    /**
+     * 用户 ID
+     */
+    @JsonProperty("userId")
+    private String userId;
+    /**
+     * 用户 ID 类型，默认值为 `user_id`，可选值为：
+     * - `user_id`: Authing 用户 ID，如 `6319a1504f3xxxxf214dd5b7`
+     * - `phone`: 用户手机号
+     * - `email`: 用户邮箱
+     * - `username`: 用户名
+     * - `external_id`: 用户在外部系统的 ID，对应 Authing 用户信息的 `externalId` 字段
+     * - `identity`: 用户的外部身份源信息，格式为 `<extIdpId>:<userIdInIdp>`，其中 `<extIdpId>` 为 Authing 身份源的 ID，`<userIdInIdp>` 为用户在外部身份源的 ID。
+     * 示例值：`62f20932716fbcc10d966ee5:ou_8bae746eac07cd2564654140d2a9ac61`。
+     * - `sync_relation`: 用户的外部身份源信息，格式为 `<provier>:<userIdInIdp>`，其中 `<provier>` 为同步身份源类型，如 wechatwork, lark；`<userIdInIdp>` 为用户在外部身份源的 ID。
+     * 示例值：`lark:ou_8bae746eac07cd2564654140d2a9ac61`。
+     *
+     */
+    @JsonProperty("userIdType")
+    private String userIdType;
+
+
+    public String getUserId() {
+        return userId;
+    }
+    public void setUserId(String userId) {
+        this.userId = userId;
+    }
+
+    public String getUserIdType() {
+        return userIdType;
+    }
+    public void setUserIdType(String userIdType) {
+        this.userIdType = userIdType;
+    }
+}
